+++ libs/compute_api/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod models;
+pub mod spec;
